renaming frontend_main to drudgeford so the app will be named drudgeford
