--- executable/build_client.txt
@@ -1,23 +1,0 @@
-To build file, run these in terminal:
-
-rm -rf build dist
-
-pyinstaller --onefile --noconsole \
-  --add-data ".:." \
-  --hidden-import pyxel \
-  --hidden-import PIL \
-  --hidden-import PIL.Image \
-  --hidden-import PIL.ImageDraw \
-  --hidden-import PIL.ImageFont \
-  frontend_main.py
-
-alternately - this runs much faster. to play it, you cd frontend_main.dist and run ./frontend_main.bin
-python3 -m nuitka --follow-imports --standalone \
-    --include-data-dir=".=." \
-    --static-libpython=no \
-    frontend_main.py
-
-python3 -m nuitka --follow-imports --standalone \
-    --include-data-dir=".=." \
-    frontend_main.py
-
